ipsec: Simplify optional XFRM policy templates
For XFRM IN and FWD policies, we have some optional templates. XFRM IN
and FWD policy templates typically act as additional filters on
decrypted packets. Making them optional simply bypasses those filters.
In that case, it therefore makes little sense to define the other
template fields.

This commit implements that change; when a template is optional, we will
null as many of its fields as possible. Optional templates will
therefore look like:

    tmpl src 0.0.0.0 dst 0.0.0.0
        proto esp spi 0x00000000 reqid 0 mode tunnel
        level use share any

Note that, as can be seen in the xfrm_state_ok() function in the kernel,
setting an SPI and a ReqID of 0 is equivalent to accepting all SPIs and
ReqIDs.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>