ipsec: Simplify XFRM IN policies
Our XFRM IN policies currently look as follows (including the change
from the previous commit).

    src 0.0.0.0/0 dst 10.242.1.0/24
        dir in priority 0
    	mark 0xd00/0xf00
    	tmpl src 10.242.0.95 dst 10.242.1.62
    	    proto esp spi 0x00000000 reqid 1 mode tunnel
    src 0.0.0.0/0 dst 10.242.1.0/24
    	dir in priority 0
    	mark 0x200/0xf00
    	tmpl src 0.0.0.0 dst 0.0.0.0
    	    proto esp spi 0x00000000 reqid 0 mode tunnel
    	    level use

This commit changes them into a single IN policy to allow everything
through:

    src 0.0.0.0/0 dst 0.0.0.0/0
        dir in priority 0
        tmpl src 0.0.0.0 dst 0.0.0.0
            proto esp spi 0x00000000 reqid 0 mode tunnel
            level use

We've always written our XFRM IN policies to try and match every
possible traffic, so this is clearly the easiest way to do that.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>